pyproject.toml
[project]
name = "tiny-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "Revanth Gundala", email = "revanth.gundala@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "torch @ https://download.pytorch.org/whl/cpu/torch-2.3.1%2Bcpu-cp312-cp312-linux_x86_64.whl",
    "torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.18.1%2Bcpu-cp312-cp312-linux_x86_64.whl",
    "vizdoom>=1.2.3,<2.0.0",
    "pandas>=2.2.0,<3.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.3.0,<11.0.0",
    "diffusers>=0.27.2,<0.28.0",
    "stable-baselines3[extra]>=2.3.0,<3.0.0",
    "transformers>=4.40.0,<5.0.0",
    "accelerate>=0.29.0,<0.30.0",
    "tqdm>=4.66.0,<5.0.0",
    "peft>=0.10.0,<0.11.0",
    "huggingface-hub>=0.22.0,<0.23.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "azure-ai-ml>=1.15.0",
    "azure-identity>=1.16.0",
    "mlflow>=2.12.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta" 